pyproject.toml
[tool.poetry]
name = "tdd"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "tdd"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.2"
djangorestframework = "^3.14.0"
django-types = "^0.16.0"
djangorestframework-types = "^0.8.0"
django-stubs-ext = "^0.5.0"
pytest-cov = "^4.0.0"
freezegun = "^1.2.2"
django-extensions = "^3.2.1"
django-filter = "^22.1"


[tool.poetry.group.dev.dependencies]
black = "^22.10.0"
flake8 = "^5.0.4"
flake8-import-order = "^0.18.1"
isort = "^5.10.1"
pytest = "^7.1.3"
pytest-django = "^4.5.2"
pre-commit = "^2.20.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "tdd.settings.testing"
markers = [
    "integration: marks tests as integration (deselect with '-m \"not integration\"')",
]
addopts = "--cov=. --cov-report term-missing:skip-covered --cov-fail-under=100 --no-cov-on-fail"

[tool.isort]
line_length = 120
lines_after_imports = 2
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
extend_skip = ["migrations"]
